Clean up some dead code and MSVC warnings.
- Dead code from experiments in SelectionBar removed.
- Many warnings about unused parameters fixed with WXUNUSED()
- Many warnings about signed / unsigned comparisons cleaned up.
- Several 'local variable declared but not used' warnings fixed.

+++ src/InterpolateAudio.cpp
@@ -112,7 +112,7 @@ void InterpolateAudio(float *buffer, const size_t len,
    const int IP =
       imin(imin(numBad * 3, 50), imax(firstBad - 1, len - (firstBad + numBad) - 1));
 
-   if (IP < 3 || IP >= N) {
+   if (IP < 3 || IP >= (int)N) {
       LinearInterpolateAudio(buffer, len, firstBad, numBad);
       return;
    }
